+++ test/commands/can/review_test.dart
@@ -111,12 +111,12 @@ void main() {
           Node(
             name: 'A',
             directory: Directory(path.join(ticketDir.path, 'A')),
-            pubspec: Pubspec('A'),
+            manifest: DartPackageManifest(pubspec: Pubspec('A')),
           ),
           Node(
             name: 'B',
             directory: Directory(path.join(ticketDir.path, 'B')),
-            pubspec: Pubspec('B'),
+            manifest: DartPackageManifest(pubspec: Pubspec('B')),
           ),
         ],
       );
@@ -168,12 +168,12 @@ void main() {
           Node(
             name: 'A',
             directory: Directory(path.join(ticketDir.path, 'A')),
-            pubspec: Pubspec('A'),
+            manifest: DartPackageManifest(pubspec: Pubspec('A')),
           ),
           Node(
             name: 'B',
             directory: Directory(path.join(ticketDir.path, 'B')),
-            pubspec: Pubspec('B'),
+            manifest: DartPackageManifest(pubspec: Pubspec('B')),
           ),
         ],
       );
@@ -234,12 +234,12 @@ void main() {
           Node(
             name: 'A',
             directory: Directory(path.join(ticketDir.path, 'A')),
-            pubspec: Pubspec('A'),
+            manifest: DartPackageManifest(pubspec: Pubspec('A')),
           ),
           Node(
             name: 'B',
             directory: Directory(path.join(ticketDir.path, 'B')),
-            pubspec: Pubspec('B'),
+            manifest: DartPackageManifest(pubspec: Pubspec('B')),
           ),
         ],
       );