+++ test/commands/do/review_test.dart
@@ -148,12 +148,12 @@ void main() {
             Node(
               name: 'A',
               directory: Directory(path.join(ticketDir.path, 'A')),
-              pubspec: Pubspec('A'),
+              manifest: DartPackageManifest(pubspec: Pubspec('A')),
             ),
             Node(
               name: 'B',
               directory: Directory(path.join(ticketDir.path, 'B')),
-              pubspec: Pubspec('B'),
+              manifest: DartPackageManifest(pubspec: Pubspec('B')),
             ),
           ],
         );
@@ -337,7 +337,7 @@ void main() {
           Node(
             name: 'A',
             directory: Directory(path.join(ticketDir.path, 'A')),
-            pubspec: Pubspec('A'),
+            manifest: DartPackageManifest(pubspec: Pubspec('A')),
           ),
         ],
       );
@@ -425,7 +425,7 @@ void main() {
             Node(
               name: 'A',
               directory: Directory(path.join(ticketDir.path, 'A')),
-              pubspec: Pubspec('A'),
+              manifest: DartPackageManifest(pubspec: Pubspec('A')),
             ),
           ],
         );
@@ -517,7 +517,7 @@ void main() {
           Node(
             name: 'A',
             directory: Directory(path.join(ticketDir.path, 'A')),
-            pubspec: Pubspec('A'),
+            manifest: DartPackageManifest(pubspec: Pubspec('A')),
           ),
         ],
       );
@@ -641,7 +641,7 @@ void main() {
           Node(
             name: 'A',
             directory: Directory(path.join(ticketDir.path, 'A')),
-            pubspec: Pubspec('A'),
+            manifest: DartPackageManifest(pubspec: Pubspec('A')),
           ),
         ],
       );
@@ -764,7 +764,7 @@ void main() {
           Node(
             name: 'A',
             directory: Directory(path.join(ticketDir.path, 'A')),
-            pubspec: Pubspec('A'),
+            manifest: DartPackageManifest(pubspec: Pubspec('A')),
           ),
         ],
       );
@@ -894,7 +894,7 @@ void main() {
             Node(
               name: 'A',
               directory: Directory(path.join(ticketDir.path, 'A')),
-              pubspec: Pubspec('A'),
+              manifest: DartPackageManifest(pubspec: Pubspec('A')),
             ),
           ],
         );
@@ -1031,7 +1031,7 @@ void main() {
             Node(
               name: 'A',
               directory: Directory(path.join(ticketDir.path, 'A')),
-              pubspec: Pubspec('A'),
+              manifest: DartPackageManifest(pubspec: Pubspec('A')),
             ),
           ],
         );
@@ -1159,7 +1159,7 @@ void main() {
             Node(
               name: 'A',
               directory: Directory(path.join(ticketDir.path, 'A')),
-              pubspec: Pubspec('A'),
+              manifest: DartPackageManifest(pubspec: Pubspec('A')),
             ),
           ],
         );
@@ -1296,7 +1296,7 @@ void main() {
             Node(
               name: 'A',
               directory: Directory(path.join(ticketDir.path, 'A')),
-              pubspec: Pubspec('A'),
+              manifest: DartPackageManifest(pubspec: Pubspec('A')),
             ),
           ],
         );